+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Releases ordered so that the most recent are displayed at the top, with the curr
 
 -   Initial plugin setup with basic repository configuration
 -   Run plugin within local WordPress Docker install during development
+-   Admin interface for configuring plugin.